# Improve load path performance