--- a/src/kernel/mppx.mli
+++ b/src/kernel/mppx.mli
@@ -1,1 +1,1 @@
-val rewrite : Mconfig.t -> Mreader.parsetree -> Mconfig.t * Mreader.parsetree
+val rewrite : Mconfig.t -> Mreader.parsetree -> Mreader.parsetree