--- a/src/kernel/mreader.mli
+++ b/src/kernel/mreader.mli
@@ -6,7 +6,6 @@ type parsetree = [
 type comment = (string * Location.t)
 
 type result = {
-  config        : Mconfig.t;
   lexer_keywords: string list;
   lexer_errors  : exn list;
   parser_errors : exn list;
@@ -42,7 +41,3 @@ val print_batch_outcome :
 
 val reconstruct_identifier:
   Mconfig.t -> Msource.t -> Lexing.position -> string Location.loc list
-
-(* Update config after parse *)
-
-val apply_directives: Mconfig.t -> parsetree -> Mconfig.t