--- a/src/kernel/mtyper.ml
+++ b/src/kernel/mtyper.ml
@@ -136,8 +136,13 @@ let type_interface config caught parsetree =
 
 let run config parsetree =
   if not (Env.check_state_consistency ()) then (
+    (* Resetting the local store will clear the load_path cache.
+       Save it now, reset the store and then restore the path. *)
+    let load_path = Load_path.get_paths () in
     Mocaml.flush_caches ();
     Local_store.reset ();
+    Load_path.reset ();
+    Load_path.init load_path;
   );
   let caught = ref [] in
   Msupport.catch_errors Mconfig.(config.ocaml.warnings) caught @@ fun () ->