This PR makes 4 changes to speed up the handling of the load path:
1. It restores the use of Directory_content_cache for the load path -- which seems to have been removed by accident in the switch to 4.12.
2. It changes where 'Mocaml.setup_config` is called to ensure that it is only called once per command.
3. It uses a hash table for the load path instead of a reference to a map.
4. It reuses the previous hash table when the load path hasn't changed and none of it's directories' contents have changed.

On a randomly chosen command in a library with a large load path, each of these changes is worth around 0.1s. #1309 is also worth about 0.1s. Between them these changes take that command's runtime from 0.7s to 0.25s. (This is for the second call to the command, the initial call to the command is more like 1.9s vs 1.7s).

I also have [a version of this branch for 4.11](https://github.com/lpw25/merlin/tree/fast-load-path-411), which is the one I actually tested.