@let-def 2021-07-29
It is not correct to setup the loadpath only in the reader.

For instance, if we switch back-and-forth between two files with different build environments without changing them, the reader will be reused so `Mocaml.setup_config` will be invoked only the first time. The second time, the typer might run in the wrong environment.

Since it is `Load_path.init` that is expensive and only the typechecker needs it, I suggest to do it only when setting up the environment for the typechecker.
 
```diff
diff --git a/src/kernel/mocaml.ml b/src/kernel/mocaml.ml
index 77f1e2aa..6b4cd38d 100644
--- a/src/kernel/mocaml.ml
+++ b/src/kernel/mocaml.ml
@@ -25,12 +25,11 @@ let is_current_state state = match !current_state with
 
 (* Build settings *)
 
-let setup_config config = (
+let setup_reader_config config = (
   assert Local_store.(is_bound ());
   let open Mconfig in
   let open Clflags in
   let ocaml = config.ocaml in
-  Load_path.init (Mconfig.build_path config);
   Env.set_unit_name (Mconfig.unitname config);
   Location.input_name  := config.query.filename;
   fast                 := ocaml.unsafe ;
@@ -46,6 +45,11 @@ let setup_config config = (
   open_modules         := ocaml.open_modules ;
 )
 
+let setup_typer_config config = (
+  setup_reader_config config;
+  Load_path.init (Mconfig.build_path config);
+)
+
 (** Switchable implementation of Oprint *)
 
 let default_out_value          = !Oprint.out_value
diff --git a/src/kernel/mocaml.mli b/src/kernel/mocaml.mli
index ffb67072..3a8fb6d5 100644
--- a/src/kernel/mocaml.mli
+++ b/src/kernel/mocaml.mli
@@ -6,7 +6,8 @@ val with_state : typer_state -> (unit -> 'a) -> 'a
 val is_current_state : typer_state -> bool
 
 (* Build settings *)
-val setup_config : Mconfig.t -> unit
+val setup_reader_config : Mconfig.t -> unit
+val setup_typer_config : Mconfig.t -> unit
 
 (* Replace Outcome printer *)
 val default_printer :
diff --git a/src/kernel/mpipeline.ml b/src/kernel/mpipeline.ml
index f6add7e0..c9700e0d 100644
--- a/src/kernel/mpipeline.ml
+++ b/src/kernel/mpipeline.ml
@@ -156,7 +156,7 @@ let process
   let reader = timed_lazy reader_time (lazy (
       let lazy source = source in
       let config = Mconfig.normalize config in
-      Mocaml.setup_config config;
+      Mocaml.setup_reader_config config;
       let result = Mreader.parse ?for_completion config source in
       result, config
     )) in
@@ -169,6 +169,7 @@ let process
     )) in
   let typer = timed_lazy typer_time (lazy (
       let lazy { Ppx. config; parsetree; _ } = ppx in
+      Mocaml.setup_typer_config config;
       let result = Mtyper.run config parsetree in
       let errors = timed_lazy error_time (lazy (Mtyper.get_errors result)) in
       { Typer. errors; result }
```

Longer term, it might be better to think about a better way to do incrementality and rewrite this `Load_path` accordingly. But it will make maintenance more difficult, the current trade-off might not be so bad.