@let-def 2021-11-22
I still had bug with it, I was able to track them down to the use of the hash table, but I am not able to explain where it is going wrong. When I stopped looking at that problem, I was thinking of rewriting parts of Load_path...
I will resume work on it on Wednesday, but the patch cannot be merged as it is.