@let-def 2021-11-25
On my fork, I rebased this branch on top of current master for testing: https://github.com/let-def/merlin/tree/fast-load-path
I will update this PR later if it works as expected.